+++ Sources/KeyboardKit/Feedback/KeyboardFeedbackContext.swift
@@ -34,12 +34,9 @@ public class KeyboardFeedbackContext: ObservableObject {
 
     // MARK: - Settings
 
-    /// A ``KeyboardFeedback/Settings`` typealias.
-    public typealias Settings = KeyboardFeedback.Settings
-
     /// Feedback-specific, auto-persisted settings.
     @Published
-    public var settings: Settings
+    public var settings: KeyboardFeedbackSettings
 
 
     // MARK: - Typealiases