+++ Sources/KeyboardKit/Themes/KeyboardThemeContext.swift
@@ -11,8 +11,8 @@ import SwiftUI
 /// This class has observable states and persistent settings
 /// for keyboard-specific themes.
 ///
-/// The class has observable auto-persisted ``settings`` for
-/// e.g. manage the current ``KeyboardTheme/Settings/theme``.
+/// This class has observable auto-persisted ``settings`` to
+/// handle the current ``KeyboardThemeSettings/theme``.
 ///
 /// Due to an observation-related issue, that caused SwiftUI
 /// to not update when a ``settings`` theme was changed, the
@@ -32,12 +32,9 @@ public class KeyboardThemeContext: ObservableObject {
 
     // MARK: - Settings
 
-    /// A ``KeyboardTheme/Settings`` typealias.
-    public typealias Settings = KeyboardTheme.Settings
-
     /// Theme-specific, auto-persisted settings.
     @Published
-    public var settings: Settings
+    public var settings: KeyboardThemeSettings
 
     /// The last theme changed date, if any.
     @Published
@@ -61,12 +58,12 @@ public extension KeyboardThemeContext {
         }
     }
 
-    /// Reset the current ``KeyboardTheme/Settings/theme``.
+    /// Reset the current ``KeyboardThemeSettings/theme``.
     func resetCurrentTheme() {
         setCurrentTheme(nil)
     }
 
-    /// Set the current ``KeyboardTheme/Settings/theme``.
+    /// Set the current ``KeyboardThemeSettings/theme``.
     func setCurrentTheme(_ theme: KeyboardTheme?) {
         settings.themeValue.value = theme
         themeChanged = Date.now.timeIntervalSince1970